+++ src/tasks/scrape-section-details.ts
